Update display selectively from add/remove
Do not update and recreate the whole display from an add or remove,
rather you can update all indices with a remove since that can happen
at any spot. We add the logic to support this. The aim is to modify the
list and the html not just the list alone while updating the entire
UI layout cards instead of just removing and inserting specific ones
but leaving the rest of the html unchanged where it does not need to be
changed. Most of the computation is simply in the background
now on a delete updating indices not on EVERY single add and delete.

This might be slightly vague so revisit the previous commit to compare
the outdated initial approach with its flaws with this new approach.