On done click note if we dealt with an edited card
Change return value back to "done" in the method. This ensures we handle
an edited card differently, and afterwards no bus where we can't break
out if there was a number and it changes back to done. Only edits to
done if there is no number directly from the edit. If we are from the
add button click though after editing it would detect a number and think
its still in edit mode so we go back to the default mode. There is
probably a better, more streamlined way to act on the edit input mode
detected than this.