Add bool for emphasis on done mode detected (non-edit mode)
Also change return value to "" default start value (no return value)
so that done can be added back the next time once the blank is detected.
Then do not convert this blank to a number since it will return as a
number which is not what I care for, and the conversion I want. This
means its not a number but rather we can go into done mode not edit mode
since a number is not detected. It's a true number conversion attempt
in js.